Adding Pathfinder cs files with 8 possible
Adding Pathfinder CS file and 15-minute talk. This is the main program that has 8 ways to solve different types of routing and pathfinding problems. Could use refactoring for cleanup and ease of reading. Break out the different challenges and resolutions. etc. I wrote this for fun and to explore different ways of solving maze/labyrinth solutions. I need to do Trémaux's and Lee algorithms. That would be cool. These solutions so far are step-by-step, have terrain and obstacles, and add randomnessto some solutions to follow a new path. Extension of this would be to map entire maze and then determine the lowest score. Add data collection for machine learning. etc.